refactor: flatten folder structure and normalize file naming
- Delete empty ghost directories (domain/config/, domain/errors/)
- Flatten single-file UI directories:
  connections/ConnectionsView.ts → ui/ConnectionsView.ts
  lookup/LookupView.ts → ui/LookupView.ts
  views/result-context-menu.ts → ui/result-context-menu.ts
- Normalize adapter file names:
  _api.ts → api-base.ts
  lm_studio.ts → lm-studio.ts
  open_router.ts → open-router.ts
- Clarify ambiguous module names:
  embedding-manager.ts → embed-orchestrator.ts
  collection-manager.ts → collection-loader.ts
- Update CLAUDE.md architecture tree
